fix all and integrate with frontend

+++ database/migrations/2021_11_26_115147_create_bills_table.php
@@ -19,7 +19,7 @@ class CreateBillsTable extends Migration
             $table->foreignId('room_id');
             $table->integer('total');
             $table->tinyInteger('month');
-            $table->tinyInteger('year');
+            $table->integer('year');
             $table->string('status')->nullable();
             $table->timestamp('pay_date')->nullable();
             $table->timestamps();
